(BKR-505) rescue SkipTest in confine_block
Prior to this, use of confine_block would result in all following code in the
test case being skipped in situations where the host list had no matches for
the confine_block.  The entire purpose of confine_block is to limit the scope
of the confine to the passed block.

Also, ensure and rescue at the method level do not require an additional
wrapping begin..end block.