(BKR-358) fix timeouts in git clone test
Prior to the PR which moved this file around, it turns out that this test was
not actually being run by jenkins.

Once it started being run, I found that the windows builds were timing out.
Digging into this a bit further, the timeout was due to the test presenting
a prompt:

```
Cloning into 'C:/cygwin64/tmp/clone_git_repo_on_test.rb.n0NGrn/puppet'...
Username for 'https://github.com':
```

and waiting forever for input. There may be a way to provide enough credential
information for windows tests to run against the git `https` protocol, but this
commit sidesteps that issue via using the `git` protocol instead of `https`.