(BKR-358) get all builds green
I'm not 100% thrilled with this approach (use git for clones on cygwin, https
everywhere else), but getting the build green is more important now, and we
can refactor or change approaches later (which seems like a matter of getting
the appropriate credentials in the appropriate places).

The root problem on Windows is that when attempting to clone a non-existent
repo on GitHub, we are prompted for username/password for https auth.