(BKR-358) streamline clone_git_repo_on acceptance test
Prior to this we were still seeing issues with getting this acceptance test
passing on all platforms.  The core issue on windows platforms is that when
attempting to clone a non-existent repository from github, credentials are
asked for (https) which appear not to be set up; and for the git protocol,
the appropriate SSH credentials also appear not to be in place.

Taking a closer look at the test, it was somewhat confused: in addition to
testing git clones, it was also trying to test the vagaries of `build_git_url`
and its environment settings (via `lookup_in_env`).  There are already spec
tests for some of this behavior, so testing it here is mostly a waste of
resources.

This commit bolsters those specs to test more of the ENV setting functionality
behind `build_git_url` (also aliased as `build_giturl`), and trims this test
down to doing a successful git clone, in hopes this will make the test happy
on all platforms.