(BKR-358) Move to an options file to specify helper path
As requested in https://github.com/puppetlabs/beaker/pull/930/#issuecomment-140548946

Here we add an options file (in `acceptance/config/acceptance-options.rb`, as
`options.rb` is excluded from git tracking via the project's `.gitignore` file).

This has the net effect of removing the `..`-heavy `$LOAD_PATH` changes, at the
expense of test runners being forced to know about and to explicitly use the
options file via:

```
% beaker --options-file acceptance/config/acceptance-options.rb # ...
```

... whenever the acceptance suite, or any part of the acceptance suite, are run.

It's not clear to me where this expectation is to be documented.  I don't see
this as a win, but am providing the requested commit to continue the discussion.

/cc @puppetlabs/beaker