(BKR-358) Mitigate the need to --option-file for full-suite runs
We introduced a required `--options-file` file in @4bd097c4 for users wishing
to run some or all of the acceptance suite.  To mitigate this overhead for
full-suite runs (the most common case), this commit provides a suite-level
file which will ensure that `$LOAD_PATH` is properly set up, in the absence
of the `--options-file` invocation.

/cc @puppetlabs/beaker

Note: since we are not on JJB for our beaker jenkins, without adding this
fix for `$LOAD_PATH` it is exceedingly difficult to run this branch through the
normal beaker jenkins acceptance testing pipeline.  Any change must be manual
to the pipeline via the web interface, and a change will either imply a
conditional in the build script (based on the presence of the options file),
or will break the other usage of the pipeline.

In the JJB world this would be cumbersome (spinning up a parallel feature
pipeline), but would be trackable via parallel `ci-job-configs` PR alongside
this branch.

/cc @puppetlabs/jjb