(BKR-522) now creates new ec2 keys per run
Before, ec2 keys would only be created if this was the first run for
a particular user/coordinator. This is a problem for F5 testing, in
which F5 hosts needed to be created with a particular key. We were
using cached keys, which weren't the ones being used in ec2.

The original solution was to delete the keys in ec2, so that they'd
be recreated as if this was the first run by a user. @justinstoller
brought up the good point that if this were to happen, certain
Beaker runs would have their keys deleted from a new F5 run. The
solution became that each Beaker run would generate its own key,
deleting it on cleanup.