(BKR-554) Install AIX packages via install repo
This commit adds support for installing AIX 'packages' via the
install repo logic. AIX does not support repository management
and can only install 'RPM' files directly. Since this logic
is contained in the install repo logic in Beaker, we can use
this to install a package that has been mirrored to the standard
repo location.

This assumes that the package is mirrored to the repo location.
In other words, that the following files are the same.
  * http://builds.delivery.puppetlabs.net/puppet-agent/1214e51d63b84a82df0c55cab99abc2a3f90a597/artifacts/aix/7.1/PC1/ppc/puppet-agent-1.2.5.49.g1214e51-1.aix7.1.ppc.rpm
  * http://builds.delivery.puppetlabs.net/puppet-agent/1214e51d63b84a82df0c55cab99abc2a3f90a597/repos/aix/7.1/PC1/ppc/puppet-agent-1.2.5.49.g1214e51-1.aix7.1.ppc.rpm

It also assumes that the desired package is not defined for AIX
in the acceptance pre-suite for the project.