(BKR-582) Colorization adjustments / CI Build
When beaker detects it is being run by Jenkins, it should adjust output
colors so that they are better handled in console output. This is done
by detecting BUILD_NUMBER environment variable, a variable that is
known to be present on Jenkins build slaves.

Adjust colorization to colors that come out better during runs. The
colors that should come up to the front are shown better when running
in the console. Additionally make it easier to see logs coming back
from the SUT with an explicit color instead of NORMAL (which is the
same setting as the current GREY).

Allow overriding log colors by passing in log_colors in options.
If a user has overridden the default colors, don't override the
colors in CI.