(BKR-574) Allow global restart_when_done => false
Previously, setting `:restart_when_done` to false on the master host, or
globally in options, did not work, because the `with_puppet_running_on`
method was checking for a truthy value.

This commit checks if the key is in the hash instead. It also preserves
the behavior where the options passed in by the test case take
precedence over the host's value.