(MAINT) Move docs to README.md, rename some files
Prior to this we were using Home.md as the root for documentation in docs/
(as a direct import from the wiki). A few files also picked up strange names at
some point in the automated cleanup process, so this fixes those, and normalizes
some file names with punctuation, etc.