(QENG-3063) Forget SSH connection IP address on reboot
Prior to this, as seen during acceptance test runs for BKR-358, on centos7
and debian7 (at least), if the DHCP lease time for the SUT was short, and a
test run was sufficiently long, a late reboot would never come back as far
as beaker was concerned.

The DHCP lease expiry would issue a new IP address for the SUT, but since
the host object would often (but not always) have the IP address cached
in `self['ip']`, when the SSH connection would be reinstantiated, it would
go to `@ip` first, before going to `@vmhostname` or `@hostname`, meaning
that connections would never connect to the new IP address.