(BKR-607) Remove tar dependency for AIX
Prior to this commit, Beaker installed -- and removed -- tar as
a dependency of the puppet-agent package on AIX. This dependency
has been removed from puppet-agent.

This commit removes the logic to install and remove the tar rpm
from AIX. It also updates the rpec test for `remove_puppet_on`
on AIX to remove the expectation that `tar` would a dependency.