(BKR-358/QENG-3063) Disable failing test to land acceptance tests
This test has proven particularly troublesome and we have a separate effort to diagnose
and deal with this issue. This has been holding up landing of these acceptance tests for
over a month and there's no reason to keep this code on a branch.

Will re-enable this test via resolving QENG-3063.