(BKR-358) Add acceptance tests for host helpers
This adds acceptance tests, along with a configuration file suitable for running
these tests in our jenkins instance, for the Beaker host helpers.

This also temporarily disables the "reboot" host acceptance test, which has been
intermittently failing, and will be re-enabled via closure of QENG-3063.

This work is detailed in the GitHub Pull Request at:

  https://github.com/puppetlabs/beaker/pull/930