(QENG-3063) Fix reboot test timeout issues on jenkins
Prior to this change, the `host_test.rb` check for rebooting systems could
sometimes fail, only on our Beaker acceptance jenkins, but in such a way that
it was clear that something was happening after the attempt to reboot started,
but no further output was available.

Exploratory work showed that when a rebooted host changes IP address, it is
possible for not only the normal long sequence of Fibonacci-fallback connect
retries to happen, but for also various low-level TCP connection failures (with
their own long timeouts) to occur.

Since output was being buffered, and not flushed, and neither of these phases
was generating a lot of output (enough to trigger a buffer flush on its own),
it was possible for more than our 10 minute jenkins output timeout to pass
without effectively changing the output buffer.

Flushing our log output overcomes the jenkins timeout problem. We do not here
address questions of long IP-address change timeouts.