(BRK-617) Print exception message for test cases
Some exceptions implement the `#inspect` method to only print the class
name or otherwise change the output to not include the error message in
question, which makes such exceptions very challenging to deubg. This
commit changes the test case exception printing method to explicitly
print the exception class and message so that there's always relevant
context available when an exception is raised.