(BKR-481) Detect SSH Connection Termination
Ubuntu 15.10 does not seem to cleanly shutdown sshd, causing the ssh
client to sit around until the next write attempt. For beaker's reboot
test this write never happens.

This change adds a keep alive and keep alive timeout to the ssh client
connection to detect when the connection is severed without
notification.

This problem is not seen on Ubuntu 15.04.