(BKR-668) dealt with on / puppet issue (details below):
if you executed this code

    on( host, puppet( 'status' ) )

on a cisco host, you would get double prefixing.  Thats because if
you pass the  method a Beaker::Command object, it would reverse
it to the command string & re-build the Command object again. This
change makes it so that if a Command object is passed, it gets used,
since the reason for creating a Command object is to pass in other
arguments youd like to be a part of execution.