(BKR-759) Update dmg url for puppet-agent install
The url path structure for osx dmg packages has changed. Also,
the `puppet-agent-latest.dmg` alias is no longer generated when
publishing packages to https://downloads.puppetlabs.com.

This commit updates the url structure to reflect the current
pattern. It also adds a method to introspect the 'latest' published
version. If a `puppet_agent_version` has not been specified and
a 'latest' version cannot be found, an exception is raised.