(MAINT) Use the upgrade install option for rpms
Images used in beaker tests might have all kinds of preinstalled
packages on them, some of which may be out of date. In our case, we
were preinstalling the puppetlabs el 7.10 release package on images,
but then running beaker's install_puppet function as a provisional step
in our tests. Since a new release package was added, this started
failing[1]. We were able to fix this easily by fixing our images, but
users depending on externally hosted images, such as vagrant boxes, may
not be so lucky.

This patch changes the -i flag of the rpm install command to -U so that
the package will be either installed or, if already present but out of
date, upgraded. This will not change functionality for users with clean
images who do not need to upgrade anything. The --replacepkgs flag that
was already present in the rpm command is not sufficient to remove
conflicting files from older versions of packages.

[1] http://logs.openstack.org/77/308277/2/check/gate-openstackci-beaker-centos-7/0736982/console.html#_2016-04-20_14_14_12_690