(BKR-779) skip_test, pending_test, fail_test don't work in with_puppet_running_on
Prior to this change, these test exits did not work in
with_puppet_running_on() as the exceptions were not bubbling up to
TestCase.
This change rescues the exceptions for these flow-controls before they
get mangled by the other rescues and sent to TestCase.  We should
probably be using throw here. But there are too many entanglements to
get them working in this PR.