(BKR-843) Use hypervisor specific logic for ip
When determining the ip address of a machine, beaker normally ssh'd into
the box and would run a command that would return the ip configured on
that box. However, this doesn't work for several cloud hypervisors, as
the public ip of the box may not match the internally configured ip
address. This change allows for hypervisor specific logic to be executed
instead of sshing into the box. Specifically, this commit changes all
ec2 instances to use the public ip address instead of the private one.