(BKR-843) Use hypervisor specific logic for ip (#1149)
* (BKR-843) Use hypervisor specific logic for ip

When determining the ip address of a machine, beaker normally ssh'd into
the box and would run a command that would return the ip configured on
that box. However, this doesn't work for several cloud hypervisors, as
the public ip of the box may not match the internally configured ip
address. This change allows for hypervisor specific logic to be executed
instead of sshing into the box. Specifically, this commit changes all
ec2 instances to use the public ip address instead of the private one.

* (BKR-843) Curl ec2 for ip in case of --no-provision

When there is no ec2 instance object created, the #get_public_ip method
needs to use curl on the box instead of using the instance object to
determine its ip. This will be the case when users add the no-provision
flag.

* (BKR-843) Add case for Windows ec2 instances for #get_public_ip