(PE-16065) Bump hocon gem to get latest beaker-answers
Currently beaker is pinning the hocon gem at a version less than 1.0
but beaker-pe requires hocon 1.0 or greater so we are essentially stuck
at an older version of beaker-answers.

This commit bumps the hocon gem in beaker to be at least 1.0 to allow us
to test with the latest beaker-answers.