(MAINT) add generic dmg and msi installers
This commit adds a generic dmg installer to the OS X package facility.
This allows either a either a local file location or a URL to be specified
to install a dmg on an OS X host. It is based on the existing `install_dmg`
method, but requires the Volume location and package name to be specified by
the caller. These values cannot be introspected without making assumptions
about the packaging.

It also adds a generic msi installer as a helper. It is based on the
`install_msi_on` helper. The additional helper requirements make it a larger
effort to port to the Windows package facility which is not pursued in this
commit. The interface of the `generic_install_msi_on` method is the same as
the `install_msi_on`, it simply removes the validations that are puppet
specific.