(QENG-3198) Mechanism for archiving SUT files
This commit adds a new host helpers method called `archive_file_from`
that is intended to be used in post-suites for grabbing arbitrary files
from the SUTs and puts them in a known location where the Jenkins
Archive Artifact plugin can find them and present them with the build
results.

The implementation is basically thin wrapper around `scp_from` that sets
up a particular directory structure with a hardcoded root directory.
This hardcoded root directory will need to be added to the Jenkins
acceptance job configurations so they can be published along with the
rest of the build results.