(BKR-894) AIX - prefer local DNS if hosts file changed
This commit adds logic to the `set_etc_hosts` method to change the
preferred DNS resolution order on AIX platforms.

By default, AIX will first use external DNS to resolve a name even
if the name has an entry in the `/etc/hosts` file. This means that
an externally resolvable hostname cannot be overridden by a local
entry in `/etc/hosts` unless the resolution order is changed. The
logic added with this commit adds the configuration to give precedence
to local DNS resolution.