(BKR-847) Fix with_host_stubbed_on
Prior to this commit, the `@teardown_procs` value was not being created
to handle the teardown referenced by calling `stub_hosts_on` within
`with_host_stubbed_on`. This commit removes the reference to the method
and adds in the desired code directly.

The `stub_hosts_on` method includes a `teardown` step to allow for the
removal of the hosts stub after the test is concluded. The `teardown`
method is incompatible with `beaker_rspec`. Therefore, modules tests using
`beaker_rspec` use the `with_host_stubbed_on` method that does not utilize
this feature.