(BKR-904) Provide A Way To Wait For A Host Reboot
Fixes in issue in Host.port_open? where an exception Errno::EHOSTUNREACH is
raised but not rescued.  Creates a wait_for_port wrapper around this function
to wait upon a specific port for a specific time.  Also allows the interval
between checks to be specified to prevent spamming the host.