(PE-17093) Solaris 10/11 pkg/pkgutil fixes
Prior to this commit Solaris 10 would blow up if pkgutil wasn't
installed. Solaris 11 was not verifying that curl/ntp were installed.

This commit adds logic to install pkgutil if it isn't found on
Solaris 11, as well as adding package checks for Solaris 11 so that
it actually verifies that ntp/curl are installed.