(BKR-916) Use dns_name for ec2 hostnames
Prior to this change, provisioning in ec2 would set each host's hostname
to the `name` attribute of the host. This change removes that behavior
and sets each host's hostname to be the public facing `dns_name`
provided by ec2 to allow traffic from outside ec2 to route correctly to
that host.