(PA-514) Ensure windows reg value & startup type
This commit adds two checks to make sure that after install our Windows agents
still have the appropriate registry value setting & location for
RememberedPuppetAgentStartupMode (which is Disabled) and that the service does
in fact have a startup type of Disabled. This is after a regression was
introduced in the MSI packaging which wrote these values to the wrong location
and then they were not retained/honored after upgrade.

The MSI flag for setting startup mode is PUPPET_AGENT_STARTUP_MODE, and it
accepts three values:

* Automatic
* Manual
* Disabled

Of these three, "Automatic" is the only on that does not correspond directly to
the equivalent value returned by `WMIC SERVICE` - for this case, `WMIC SERVICE`
returns "Auto" and not "Automatic. "Manual" and "Disabled" are both returned
as-is. For this reason, Automatic is treated as a special case in this commit.

In querying for the puppet service using WMI, a Wildcard match ('%') is prepend
to Puppet to ensure we can match the PE 3.8.x service name of "PE-Puppet".

This location (beaker) is chosen for this test because the puppet
agent testing pipeline does not include its own testing for package-based
upgrades, and at this time all systems that install MSIs will use this util. At
such time as package acceptance testing of upgrades is added to puppet-agent,
these checks should be moved there.

Paired with: Sean McDonald

Signed-off-by: Moses Mendoza moses@puppet.com