(BKR-911) Add OSX Sierra (10.12) support
Add the osx 10.12 entry to the platform config, and update the
ssh_permit_user_environment function to also use /private/etc/sshd_config
instead of /etc/sshd_config for osx-10.12.

osx-10.11 and 10.12 both need that location to be updated (with the addition
of /private) in order for the ssh environment to work correctly.