(BKR-952) Add puppet_version_comparison
The `version_is_less` method is widely used for comparing PE and Puppet
version across all of the acceptance tests. But only being able to
perform a less than comparison is limiting and hard to read since
you must check `!version_is_less` to determine if versions are
greater than.

This commit adds a `puppet_version_comparison` method that accepts
a version `a`, a comparison operator `cmp` and a version `b`.
It breaks up `a` and `b` into all of their various version parts that
we expect and compares each value in order using `cmp`.