(BKR-954) Add Ability to Disable Openstack Volume Management
Prior to this commit beaker openstack support assumed that the
`volumes` service was provisioned.  This service is not always
available in an Openstack deployment.  This commit gives you the
option to disable that functionality.  If you do not disable it
on an instance that does not publish this functionality, even if
you don't configure the volumes setting in `CONFIG`, beaker runs
will fail.