(MAINT) Allow no hosts option to be passed
In previous versions (such as 2.14.1, which was previously
used to run our ec2 zombie killer), beaker could be run
without a hosts file argument. This would allow beaker test
logic to be run without needing to stand up SUTs. During the
time where beaker-hostgenerator support was added, this
ability was removed. This change allows us to run without
the hosts file argument, as was previously done.