(MAINT) Update CLI hosts option help text
In #1266, we added the ability to have the  CLI
argument be omitted for a beaker run. This was restoring
old behavior, but there was no guide on what values could
be passed. This change sets the CLI help text to explain
the possibilities to the user, clearing up the ways we
intend the  argument to work.