(BKR-971) Ensure scp errors close ssh connections
Previous to this commit, beaker scp errors would leave the ssh
connection in a dangling state because of Exceptions that would bubble
up without closing out the ssh connection. This change ensures that when
errors are raised, the #close method is called. While the exceptions are
swallowed, the #close methods called end up raising errors from the
state of the connection object, thus not requiring that the original
asserts be raised.