(MAINT) clarify puppet_agent_version parameter
Before, it was unclear why `puppet_agent_version` was
being used. It's reason for existance is still a little
unclear, but we were at least able to clarify how it's
used from the client side. Added some docs to explain
that, and a note to clean this up for future versions.