(BKR-1006) Close connection immediately when rebooting Windows
When using `Windows::Exec::reboot`, the SSH connection is expected to
fail. However, it can take a long time to do so (greater than 30 minutes
in some cases). This should be faster, as we know the connection will
fail.

Add a new option to `exec`, `:reset_connection`, that immediately closes
the connection after sending a message. Use it when performing reboot on
a Windows host.