(BKR-1016) resolve stub forge from SUT
This commit modifies the implementation of the `stub_forge_on` and
`with_forge_stubbed_on` methods to resolve the `forge_host` from
the host on which the targeted `forge_host` will be accessed.

This is done to address the case where the SUT and the Beaker
runner have different resolution of the given `forge_host` name.