(BKR-1051) Add retry loop for aws key check
The kb.exists? check fails intermittently; it one of the failures, manual
inspection shows that the key pair was in fact created. This heavily implies
that there's a race at work; where the key pair doesn't show as existing until
some time after the initial creation.

Add a retry loop around the kb.exists? check to attempt to work around this
issue.