(maint) stub out wait_for_host_in_dashboard
wait_for_host_in_dashboard is no longer needed with PE and is causing
ihssues with attempts to turn off node-checkin by default in classifier.

This removes the check that validates a node is seen by the classifier.

It shouldn't be needed anyway as a puppet run has already been done by
the time this is run.