(BKR-1046) Use version for pdb status in legacy
This commit adds logic based on `pe_ver` to use the PuppetDB `version`
API endpoint when the version includes PuppetDB version less than 4.0
which do not include the `status` API endpoint.

Prior to this commit, the status check on PE versions less than 2016.1.0
would fail because the `status` endpoint was not available.