(MODULES-4594) vmpooler hypervisor: add disks
In order to set up a test environment for Red Hat Satellite, we need a
disk that's larger than the default vmpooler allotment (about 5 GB free
vs the 1.5GB that are free on a redhat host after all is said and done).
This adds a step to the provision function in the vmpooler hypervisor to
add a disk/s, x GB in size, to any host in the host configuration file. It is
then up to the tests to partition, mount, expand, or whatever they
decide to do with it/them.