(BKR-1091) vCloud datacenter selection support
prior to this commit you couldn't specify the datacenter to search for a
given folder.  The `find_datacenter` method will return the first found
datacenter, and in the case that you have linked vcenters this falls
apart.  This commit adds the ability for `datacenter` to be included in
@options.  Related, folder search now respects datacenter as well as
uses the traverse method.  Prior to this commit some folder searches
were not possible (for example a base level because it didn't contain a
`/` which was being split on)