(BKR-1124) Pass info text to summary logger to print on file
This patch passes test summary info to summary_logger which then prints on the .txt file.
The summary_logger instance prints test summary messages to STDOUT and on the .txt file. But when print_test_result  method is called it used @logger instance to print output to STDOUT only and there was no message on the .txt file. Now print_test_result method returns a string to summary_logger that prints to both STDOUT and to the .txt file. The spec tests are modified so there are no expectations to the @logger instance rather runs similar tests on the string output of print_test_result.