(BKR-871) Remove cygwin installation from package installation
Previous to this commit, beaker would ensure that the cygwin executable
was available prior to package installation and download the executable
if it was not found. Later versions of cygwin fixed a flaw in permissions
for the executable, which broke the way beaker ensured and downloaded
the cygwin executable.

This change removes the notion of checking for cygwin at all. Rather
than conflate the installation of cygwin with package installation,
beaker should separate those concerns. Cygwin should be downloaded and
installed separately from package installation.

Also included in this PR is an optimization for determining the the
architecture of the host; rather than spend time introspecting the
system, a host should just examine its own platform string to determine
the architecture.