(BKR-1154) place prepend cmds before env vars in commands
Previously, beaker put environment variables first, followed by
the prepended command. This can break in a couple of ways:

* If the prepended command is a shell builtin, it will not work
  with `env`.
* Even if it could, the environment variables would be applied
  to the prepended command, not to the actual command.

An example of such a broken command is:
`env LD_LIBRARY_PATH=/foo source /etc/profile; puppet --version`

Instead, beaker should put the prepended commands first. This allows
it to work correctly if it's a shell builtin, and ensures that the
environment variables are properly passed to the intended command.

An example of the fixed command is:
`source /etc/profile; env LD_LIBRARY_PATH=foo puppet --version`