(BKR-1155) Allow configurable ssh connection preference
This patch adds a new API method for hypervisors called connections_preference.
Hypervisor authors can overwrite this method to return an array of SSH
connection method preference of their choice when that hypervisor is used. This
option is then added to host_hash.

In host.rb file, we are sending ssh_connection.rb file an object of
:ssh_connection_pereference as an option. And we loop through each of the
element of the array in connect method of ssh_connection file.

The overall idea here is to create an API endpoint for hypervisor authors to set
their connection preference. As of now we dont want end user to set this in
their hosts config as there are no use case for it.

[skip ci]