(BKR-1184) Use REXML to remove nokogiri from beaker
One functionality change too. Before this we were using export key as an xml
attribute and export value as the value for that attribute. For example:
if export key is middle earth and value is Elves we were creating xml like:

<element middle earth="Elves"> </element>

This is invalid XML. You cannot have whitespaces in attribute name.
So we are now replacing spaces with _. So now the xml would look like:

<element middle_earth="Elves"> </element>