(CPR-439) auto import gpg keys with zypper (#1449)
When enabling repos on sles, zypper doesn't seem to acknowledge the
gpgkey entry in the repo file. Generally this will cause yum to auto
import the specified gpg key, and everything is hunky dory. With zypper,
it fetches the gpg key from the repo itself (found in
repodata/repomd.xml.key). This will allow zypper to install the package
and verify the key without failing, assuming the key provided in the
repo is the correct key.