(BKR-1168) Fix the `beaker --version` not each_keys method found
The issue was happing because beaker will run the function that needs
the HOSTS arrays, which is defined as null when running `beaker --version`,
therefore ruby thrown noMethod for nilClass.

This commit add an expression to return the code when using '--version'
option, instead of continuing running the beaker code.