Allow nodeset YAML files to contain ERB
This allows users to embed ERB content into their nodeset YAML files.
This is particularly important when doing Docker testing on a shared
environment since the nodes cannot have overlapping names.

After thinking about a lot of different potential solutions, this one
appeared to be the most flexible and allow users to do what they needed
to in a language that is understood to Puppet users.

If no ERB content is present, there will be no change noticed by
current users.