(maint) Allow fine-grained, structured timing for the tests
Previously, Beaker stored information for the duration of a single
test case. However, there was no way to obtain the duration of each
piece of the test (e.g. how long a specific step or sub-test took).
This commit adds this capability -- refer to the comments above the
"timed_node" instance in the TestCase class for more details on how
it works/is stored. Note that the timing specific information is
written to a single file in the same directory as the existing
summary.txt files under the basename #{suite}-times.txt.

This commit allows for the possibility of more fine-grained metrics
tracking (e.g. which step in a test failed, which sub-test failed,
etc.) in the future by extending the timed_node object to instead
represent an AST for the test case.