(maint) Update the Ruby install of the pre-suite to force install bundler
When a new VM is checked out from the vmpooler, bundler is already installed.
This bundler installation causes a conflict when the beaker acceptance tests
attempt to install bundler, which in turn, causes the acceptance tests to fail.
This change adds a --force flag to the bundler install, ignoring the conflict