(BKR-1268) Load project options from .beaker.yml
If .beaker.yml is present in the current working directory, then load
the project's options. The overall precedence from lowest to highest is:

    defaults
    .beaker.yml
    ~/.beaker/subcommand_options.yaml
    ./.beaker/subcommand_options.yaml
    etc

The project options (.beaker.yml) are essentially defaults for the
project, so they come after the hard-coded beaker defaults, but before
the subcommand options.

We always load the project options for all subcommands. This is unlike
the subcommand options, because those need to be ignored when starting a
new beaker subcommand session (hence we skip loading when the subcommand
is help or init).

It is also possible for the project options to specify the options_file
to load so that it doesn't need to be specified on the command line.

Note this fixes a spec test that was stubbing part of the parse_args
method that it was trying to test.