(BKR-487) Remove AWS ojbects from host_preserved
This is a medium term fix for BKR-487.
The issue is that AWS objects get written to the host_preserved.yml file, but they can't be read in
because the AWS gem isn't loaded at read time.  And apparently even if it is... the initiate returns nil
which psych doens't like... (see the ticket for details)

This change is to convert those AWS objects to hashes just before writing the yml file.
This is a pretty poor fix overall.  But it will allow loading the host_preserved.yml file and running things
on the aws box.  But the later runs won't be able to kill the box or use any of the aws instance object methods...
But can't do that right now anyway, so it is a step forward. :)