(BKR-1439) Make Solaris 11 use OpenCSW for package installation
Previously Solaris 11 used a built-in repo with the 'pkg' command to
install third-party packages, and the git package provided in that repo
does not support TLS 1.2 and is no longer compatible with GitHub https:
repos. This updates Solaris 11 to behave the same as Solaris 10, which
uses the OpenCSW repo which offers a wide variety of packages, including
a git package which is compatible with GitHub https: clones.