(maint) Ensure options are created
Previously, the CLI's options hash was only created if its parse_options
method was called, but that method is often stubbed out. Initialize
options to an empty hash to ensure the CLI is not in an half-consistent
state.

Add a `cli` reader to the Subcommand so that we don't use
instance_variable_get and *_any_instance_of(Beaker::CLI) in our tests.