(BKR-1466) Accept comma-separated tests for exec subcommand
Beaker's exec subcommand will now accept a comma-separated list of
file/directories. Each directory will be recursively globbed. All other
suites, e.g. pre-suite, will be skipped.

This commit preserves the ability to specify a comma-separated list of
suites, e.g. beaker exec pre-suite,tests. Beaker will skip any suite not
listed.  Note the tests to run must have have been specified in the
options file or specified on the command line:

    beaker exec pre-suite --pre-suite step1.rb

It's kind of strange that the exec subcommand takes either a list of
testcases to execute or the names of suites to execute, because
specifying `beaker exec tests` is ambiguous. It could be "execute all of
the testcases in the tests directory", or "execute the 'tests' suite".
This commit preserves the ambiguity, but it means you can't specify a
suite name and a testcase name, e.g. pre-suite,tests/step1.rb.